Finished creating a fade test. This will allow us to fade any region on the canvas into a background image.
Began an "end point" finding algorith. This will be able to find the start and
end points of a given letter for any non-cursive font (I started with the
letter "e")

I will likely put this on hold and hard code the regions in. Once we are passed
the first demo, I will continue putting the algorithm together.

TODO: Find a way to import files into the main .pde. For some reason, the
import function doesn't seem to work....I'm considering using an outside
consolidator that will compile all of the .pde files into one giant file.